Made Bat.cs abstract. Now any new bat can implement Bat.cs and inherit all of the shooting/meter behavior.

+++ Content/UI/ShotPowerMeter.cs
@@ -6,10 +6,7 @@ using Terraria.ModLoader;
 using Terraria.UI;
 using Terraria.GameContent;
 using System.Collections.Generic;
-using Terraria.Localization;
 using Baseball.Content.Items.Weapons;
-using Steamworks;
-using System;
 using Baseball.Common.Players;
 
 namespace Baseball.Content.UI
@@ -86,7 +83,7 @@ namespace Baseball.Content.UI
             if(Main.LocalPlayer.HeldItem.ModItem is not Bat) return; // only need to do anything if we're holding a bat
 
             var modPlayer = Main.LocalPlayer.GetModPlayer<ShotPowerPlayer>();
-            powerText.SetText((modPlayer.power * 100).ToString());
+            powerText.SetText(((int)(modPlayer.power * 100)).ToString());
 
             if(modPlayer.isCalibratingPower) modPlayer.CalibratePower(gameTime); // if we are calibrating power, tell the modplayer to do that. can't do that in the modplayer because there's no Update() there
